+++ exe-unit/Cargo.toml
@@ -44,10 +44,10 @@ ya-utils-networking = { version = "0.2", default-features = false, features = [
     "dns",
     "vpn",
 ] }
+ya-utils-process = "0.3"
 ya-packet-trace = { git = "https://github.com/golemfactory/ya-packet-trace" }
 ya-gsb-http-proxy = { path = "../exe-unit/components/gsb-http-proxy" }
 ya-counters = { path = "../exe-unit/components/counters", features = ["os"] }
-ya-process = { path = "../exe-unit/components/process" }
 
 actix = { version = "0.13", default-features = false }
 actix-rt = "2.7"
