--- exe-unit/components/process/Cargo.toml
@@ -1,41 +1,0 @@
-[package]
-name = "ya-process"
-version = "0.1.0"
-edition = "2021"
-description = "Runtime process handling utilities"
-
-[dependencies]
-ya-agreement-utils = { workspace = true }
-
-anyhow = "1.0"
-actix = "0.13"
-actix-rt = "2.7"
-chrono = "0.4"
-log = "0.4"
-serde = { version = "^1.0", features = ["derive"] }
-serde_json = "1.0"
-serde_yaml = "0.9"
-thiserror = "1.0"
-tokio = { version = "1", features = [
-    "process",
-    "signal",
-    "time",
-    "net",
-    "rt-multi-thread",
-] }
-lazy_static = { version = "1.4.0" }
-
-bytes = { version = "1", optional = true }
-futures = { version = "0.3" }
-
-[target.'cfg(target_family = "unix")'.dependencies]
-nix = { version = "0.22.0" }
-
-[target.'cfg(target_os = "macos")'.dependencies]
-libproc = { version = "0.7" }
-
-[target.'cfg(target_os = "windows")'.dependencies]
-winapi = { version = "0.3.8", features = [
-    "jobapi2",
-    "processthreadsapi",
-] }
