+++ golem_cli/Cargo.toml
@@ -11,7 +11,7 @@ ya-compile-time-utils = "0.2"
 ya-core-model = { version = "^0.9", features=["payment", "version"] }
 ya-provider = "0.3"
 ya-utils-path = "0.1.0"
-ya-utils-process = { version = "0.2", features = ["lock"] }
+ya-utils-process = { version = "0.3", features = ["lock"] }
 
 actix-rt="2.7"
 ansi_term="0.12.1"
