+++ utils/process/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
 name = "ya-utils-process"
-version = "0.2.0"
+version = "0.3.0"
 authors = ["Golem Factory <contact@golem.network>"]
-edition = "2018"
+edition = "2021"
 homepage = "https://github.com/golemfactory/yagna"
 repository = "https://github.com/golemfactory/yagna"
 license = "LGPL-3.0"
@@ -15,19 +15,34 @@ lock = ["fs2"]
 
 [dependencies]
 actix = { version = "0.13", default-features = false }
+# actix-rt = "2.7"
 anyhow = "1.0"
+# bytes = { version = "1", optional = true }
+# chrono = "0.4"
 derive_more = { workspace = true }
 futures = "0.3"
 futures-util = "0.3.4"
+lazy_static = "1.4"
 libc = "0.2"
 log = "0.4"
 shared_child = "0.3.4"
 tokio = { version = "1", features = ["process", "signal"] }
+thiserror = "1.0"
+# serde = { version = "^1.0", features = ["derive"] }
+# serde_json = "1.0"
+# serde_yaml = "0.9"
 
 fs2 = { version = "0.4.3", optional = true }
 
 [target.'cfg(target_family = "unix")'.dependencies]
-nix = "0.22.0"
+nix = "0.22"
 
 [target.'cfg(target_family = "windows")'.dependencies]
-winapi = "0.3"
+winapi = { version = "0.3", features = [
+    "jobapi2",
+    "processthreadsapi",
+] }
+
+
+[target.'cfg(target_os = "macos")'.dependencies]
+libproc = { version = "0.7" }
